scene: Introduce render graph texture type
This is the prerequisite for introducing render nodes and layer
textures.

The KrkTexture class provides an API similar to the QSGTexture's API.

The KrkTexture is effectively split in two parts - the generic render
graph texture and the platform-specific texture, e.g. OpenGL texture or
QImage. The split is necessary to allow implementing atlas textures.

The decoration item puts decoration borders in a texture atlas. This is
not feasible currently as every decoration part will need a separate
geometry node and we don't have paint nodes yet. It'll be added later.

Note that a scene graph and a render graph are different. A scene graph
contains items such as surface items. A render graph specifies how the
items must be rendered. For now, there is no render graph, but it will
contain things such as texture render nodes, transform nodes, etc.