Test DrmPipelines for outputs
Not all combinations of connectors, crtcs and planes will work
on all hardware, so we need to test the pipelines before using
them. If no combinations are found using the available resources
then we turn all the outputs off first and search for a working
configuration of all connectors, crtcs and planes.

BUG: 433107
CCBUG: 435265