Record GPU render time
If the GPU is moderately loaded, the GPU render time can be higher than
the CPU render time. Currently, our frame scheduling algorithm doesn't
account for that. This means that there may be frame drops when playing
video games.

With this change, opengl timer query objects will be inserted before and
after finishing a compositing cycle. On a page flip, the final render
time will be computed based on the timer queries and the cpu render time.